Edited SetPropertyAction to use IOObject instead of Component and moved setProp to IOObject
