+++ src/app/core/actions/SetPropertyAction.ts
@@ -1,36 +1,36 @@
 import {Action} from "core/actions/Action";
 
-import {Component} from "core/models";
+import {IOObject} from "core/models";
 
 import {Prop} from "core/models/PropInfo";
 
 
 export class SetPropertyAction implements Action {
-    private component: Component;
+    private obj: IOObject;
 
     private propKey: string;
 
     private initialProp: Prop;
     private targetProp: Prop;
 
-    public constructor(component: Component, key: string, prop: Prop) {
-        if (!component.hasProp(key))
-            throw new Error(`Cannot find property ${key} in ${component.getName()}!`);
+    public constructor(obj: IOObject, key: string, prop: Prop) {
+        if (!obj.hasProp(key))
+            throw new Error(`Cannot find property ${key} in ${obj.getName()}!`);
 
-        this.component = component;
+        this.obj = obj;
         this.propKey = key;
-        this.initialProp = component.getProp(key);
+        this.initialProp = obj.getProp(key);
         this.targetProp = prop;
     }
 
     public execute(): Action {
-        this.component.setProp(this.propKey, this.targetProp);
+        this.obj.setProp(this.propKey, this.targetProp);
 
         return this;
     }
 
     public undo(): Action {
-        this.component.setProp(this.propKey, this.initialProp);
+        this.obj.setProp(this.propKey, this.initialProp);
 
         return this;
     }
